Update filenames, namespaces, and NuGet packages
