+++ src/dotnet/Core/Core.csproj
@@ -16,7 +16,7 @@
 		<PackageReference Include="Microsoft.Azure.Cosmos" Version="3.37.0" />
 		<PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
-		<PackageReference Include="Polly.Core" Version="8.2.0" />
+		<PackageReference Include="Polly.Core" Version="8.3.1" />
 	</ItemGroup>
 
   <ItemGroup>
