+++ src/dotnet/Management/Management.csproj
@@ -14,7 +14,7 @@
     <PackageReference Include="Azure.Data.AppConfiguration" Version="1.3.0" />
     <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.37.0" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
-    <PackageReference Include="Polly.Core" Version="8.2.0" />
+    <PackageReference Include="Polly.Core" Version="8.3.1" />
   </ItemGroup>
 
   <ItemGroup>
