+++ src/dotnet/OrchestrationAPI/OrchestrationAPI.csproj
@@ -12,9 +12,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-	  <PackageReference Include="Asp.Versioning.Http" Version="7.1.0" />
-	  <PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" Version="7.1.0" />
-	  <PackageReference Include="Azure.Identity" Version="1.10.4" />
+    <PackageReference Include="Asp.Versioning.Http" Version="8.1.0" />
+    <PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" Version="8.1.0" />
+    <PackageReference Include="Microsoft.Identity.Web" Version="2.17.1" />
 	  <PackageReference Include="Azure.Monitor.OpenTelemetry.AspNetCore" Version="1.1.0" />
 	  <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.21.0" />
 	  <PackageReference Include="Microsoft.ApplicationInsights.Profiler.AspNetCore" Version="2.5.3" />
