+++ tests/dotnet/Orchestration.Tests/Orchestration/OrchestrationBaseTests.cs
@@ -3,8 +3,10 @@ using FoundationaLLM.Common.Models.Hubs;
 using FoundationaLLM.Common.Models.Orchestration;
 using FoundationaLLM.Orchestration.Core.Interfaces;
 using FoundationaLLM.Orchestration.Core.Orchestration;
+using NSubstitute;
+using Xunit;
 
-namespace FoundationaLLM.AgentFactory.Tests.Orchestration
+namespace FoundationaLLM.Orchestration.Tests.Orchestration
 {
     public class OrchestrationBaseTest
     {
