+++ tests/dotnet/Orchestration.Tests/Orchestration/OrchestrationBuilderTests.cs
@@ -7,8 +7,11 @@ using Microsoft.Extensions.Configuration;
 using System.Reflection;
 using FoundationaLLM.Orchestration.Core.Interfaces;
 using FoundationaLLM.Orchestration.Core.Orchestration;
+using Microsoft.Extensions.Logging;
+using NSubstitute;
+using Xunit;
 
-namespace FoundationaLLM.AgentFactory.Tests.Orchestration
+namespace FoundationaLLM.Orchestration.Tests.Orchestration
 {
     public class OrchestrationBuilderTests
     {
